refactored code for bookings routes and change bookings.js to auth.js to avoid confusion and renamed booking to bookings for consistency with other files
