+++ backend/package.json
@@ -14,7 +14,7 @@
     "express": "^4.21.2",
     "express-oauth2-jwt-bearer": "^1.6.0",
     "mongodb": "^6.12.0",
-    "mongoose": "^8.9.5",
+    "mongoose": "^8.10.1",
     "nodemon": "^3.1.9"
   }
 }
