+++ backend/server.js
@@ -4,7 +4,7 @@ const express = require("express");
 const mongoose = require("mongoose");
 const webpageRoutes = require("./routes/webpages");
 const reviewRoutes = require("./routes/reviews");
-const bookingsRoutes = require("./routes/booking");
+const bookingsRoutes = require("./routes/bookings");
 const { errorHandler } = require("./middleware/error.js");
 
 //invoke express app
@@ -22,7 +22,7 @@ app.use((req, res, next) => {
 // page routes
 app.use("/api/pages/", webpageRoutes);
 app.use("/api/pages/", reviewRoutes); // review routes, will change routing after
-app.use(`/api/bookings/`, bookingsRoutes);
+app.use("/api/bookings", bookingsRoutes);
 
 // connect to db
 mongoose
